fix: Config with just ignores should check ignores
When a config has `ignores` and not `files`, plus at least one other
key, it should match all files that don't match `ignores` and the other
keys should be merged into the resulting config.

Prior to this, such a config was always merged in regardless if the
filename matched `ignores`; after this change, it correctly merges in
only when `ignores` matches.

Refs eslint/eslint#17103